feat: Making outdated version warning more visible (#301)
I got trapped yesterday with @novotnyj by using the old version of Apify
CLI, which creates a `./apify_storage` directory, but crawlee uses a`
./storage` directory. The same way as @lhotanok today, see:
https://apifier.slack.com/archives/CD0SF6KD4/p1669385459550849

So I am proposing 3 changes here:
- Show warning using an init [Oclif hook](https://oclif.io/docs/hooks)
instead of a prerun hook - this way, it gets always displayed - even
with `--help` or when using non-existing command
- Showing it all the time. Now it's shown only randomly on about every
10th call.
- Styling it differently with some spacing between warning and command,
see:

BEFORE:

<img width="909" alt="obrazek"
src="https://user-images.githubusercontent.com/594801/204019385-023137c1-b2c0-4a3f-a031-a0fff4280890.png">

AFTER:

<img width="908" alt="obrazek"
src="https://user-images.githubusercontent.com/594801/204019101-5996b087-faf5-4838-b04e-fcdb8c7850fc.png">

+++ test/commands/call.js
@@ -25,7 +25,7 @@ const EXPECTED_INPUT_CONTENT_TYPE = 'application/json';
  * @returns {Object}
  */
 const waitForBuildToFinish = async (client, buildId) => {
-    while (true) {
+    while (true) { // eslint-disable-line
         const build = await client.build(buildId).get();
         if (build.status !== ACT_JOB_STATUSES.RUNNING) return build;
         await new Promise((resolve) => setTimeout(resolve, 2500));
@@ -51,6 +51,7 @@ describe('apify call', () => {
     before(async function () {
         if (fs.existsSync(GLOBAL_CONFIGS_FOLDER)) {
             // Skip tests if user used CLI on local, it can break local environment!
+            console.warn(`Test was skipped as directory ${GLOBAL_CONFIGS_FOLDER} exists!`);
             this.skip();
             return;
         }
@@ -97,6 +98,7 @@ describe('apify call', () => {
     });
 
     after(async () => {
+        if (!actorId) return; // Test was skipped.
         await testUserClient.actor(actorId).delete();
         process.chdir('../');
         if (fs.existsSync(ACTOR_NAME)) await rimrafPromised(ACTOR_NAME);
